Trend_analyser/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=0.21.0
numpy>=1.21.0